Better move this to its own class, creating MachineBlockBreakHandler
